Fixed continious reset

+++ InterchipDMA.h
@@ -57,8 +57,8 @@ typedef struct _AMData {
 
 #if PATH_MANAGER
 typedef struct _GPSData {
-    double latitude;  //
-    double longitude; //
+    long double latitude;  //
+    long double longitude; //
     float time;     //4 Bytes
     float speed;
     int altitude;
